Reformat + changed function random.rs
